Move registry into own package
