+++ bundles/org.palladiosimulator.textual.tpcm/src/org/palladiosimulator/textual/tpcm/registry/ChildPropertyMapping.xtend
@@ -1,4 +1,4 @@
-package org.palladiosimulator.textual.tpcm.generator
+package org.palladiosimulator.textual.tpcm.registry
 
 interface ChildPropertyMapping<S, OT> {
     def void run(S element, OT mappedParent, GeneratorTransformationRegistry registry)
