+++ bundles/org.palladiosimulator.textual.tpcm/src/org/palladiosimulator/textual/tpcm/registry/RegistrationConfigurerImpl.xtend
@@ -1,4 +1,4 @@
-package org.palladiosimulator.textual.tpcm.generator
+package org.palladiosimulator.textual.tpcm.registry
 
 import java.util.function.Function
 import java.util.List
