+++ bundles/org.palladiosimulator.textual.tpcm/src/org/palladiosimulator/textual/tpcm/registry/TransformationRegistryConfigurer.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.textual.tpcm.generator;
+package org.palladiosimulator.textual.tpcm.registry;
 
 public interface TransformationRegistryConfigurer {
 	void configure(GeneratorTransformationRegistry registry);
