+++ bundles/org.palladiosimulator.textual.tpcm/src/org/palladiosimulator/textual/tpcm/registry/ValueApplierConfigurer.xtend
@@ -1,4 +1,4 @@
-package org.palladiosimulator.textual.tpcm.generator
+package org.palladiosimulator.textual.tpcm.registry
 
 import java.util.function.BiConsumer
